% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{demo_sce}
\alias{demo_sce}
\title{Demonstration Data Sets}
\usage{
demo_sce(name = c("pbmc3k"))
}
\arguments{
\item{name}{Data set identifier. Only \code{pbmc3k} is available at the moment.}
}
\value{
A \code{SingleCellExperiment} object.
}
\description{
Demonstration Data Sets
}
\examples{
# This example only works on the Docker container for the workshop
if (interactive()) {
    sce <- demo_sce("pbmc3k")
}
}
