% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{get_bib_file}
\alias{get_bib_file}
\title{Bibliography File Path}
\usage{
get_bib_file()
}
\value{
A character value representing the path to the bibliography
}
\description{
Bibliography File Path
}
\examples{
get_bib_file()
}
