% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEWorkshopEuroBioc2020-package.R
\docType{package}
\name{iSEEWorkshopEuroBioc2020-package}
\alias{iSEEWorkshopEuroBioc2020}
\alias{iSEEWorkshopEuroBioc2020-package}
\title{iSEEWorkshopEuroBioc2020: Interactive Visualization of SummarizedExperiment Objects with iSEE}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This workshop covers basic usage of iSEE for 
    interactive visualization of SummarizedExperiment objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://iSEE.github.io/iSEEWorkshopEuroBioc2020}
  \item Report bugs at \url{https://github.com/iSEE/iSEEWorkshopEuroBioc2020/issues}
}

}
\author{
\strong{Maintainer}: Kevin Rue-Albrecht \email{kevin.rue-albrecht@imm.ox.ac.uk} (\href{https://orcid.org/0000-0003-3899-3872}{ORCID})

Authors:
\itemize{
  \item Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})
  \item Charlotte Soneson \email{charlottesoneson@gmail.com} (\href{https://orcid.org/0000-0003-3833-2169}{ORCID})
  \item Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com} (\href{https://orcid.org/0000-0002-3564-4813}{ORCID})
}

}
\keyword{internal}
