% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{load_demo_sce}
\alias{load_demo_sce}
\alias{path_demo_sce}
\alias{cache_demo_sce}
\title{Demonstration Data Sets}
\usage{
load_demo_sce(name = c("pbmc3k"))

path_demo_sce(name)

cache_demo_sce(x, name)
}
\arguments{
\item{name}{Data set identifier.}

\item{x}{A \code{SingleCellExperiment} object.}
}
\value{
\code{load_demo_sce()} returns a \code{SingleCellExperiment} object.

\code{path_demo_sce()} returns the path to the cached RDS file.

\code{cache_demo_sce()} returns \code{NULL}; it is called for its side effect of writing a cache file.
}
\description{
Demonstration Data Sets
}
\examples{
library(SingleCellExperiment)

x <- SingleCellExperiment()
cache_demo_sce(x, "empty")

sce <- load_demo_sce("empty")

path_demo_sce("empty")
}
